Aurora backed swapper
Summary: Aurora needs to integrate swapping into the filesystem in order for checkpointing and restoring to function correctly; swap partitions and associated metadata are not persistent, so any paged out data are currently lost on a crash. Create a custom SLS swapper that overrides the default FreeBSD swapper.

Test Plan:
The new lazy restore test passes:

Build: FreeBSD 12.1-RELEASE FASTDBG
Test started: Tue Nov 3 22:23:34 EST 2020
Test                                       Status      Time
-----------------------------------------------------------
001-load-unload.sh                       [ success ]
002-load-mount-unload.sh                 [ success ]
003-mount-unmount.sh                     [ success ]
004-load-double-sls.sh                   [ success ]
005-unload-double-sls.sh                 [ success ]
006-load-double-slos.sh                  [ success ]
007-unload-double-slos.sh                [ success ]
008-mount-double.sh                      [ success ]
009-unload-slos-before-sls.sh            [ success ]
010-unmount-before-sls.sh                [ success ]
02-simple-file-io.sh                     [ success ]
20-simple-mem-ckpt.sh                    [ success ]
204-print-stdout.sh                      [ success ]
205-array-traversal.sh                   [ success ]
21-simple-osd-ckpt.sh                    [ success ]
221-lazy-osd-ckpt.sh                     [ success ]

Reviewers: ali, tavianator, ryan

Reviewed By: ali, ryan

Tags: #aurora

Differential Revision: https://review.rcs.uwaterloo.ca/D383

+++ tests/02-simple-file-io.sh
@@ -1,8 +1,13 @@
-#!/bin/sh -e
+#!/bin/sh
 
 . aurora
 
 aursetup
+if [ $? -ne 0 ]; then
+    echo "Failed to set up Aurora"
+    exit 1
+fi
+
 echo foo > $MOUNTPT/foo
 rm $MOUNTPT/foo
 
@@ -15,5 +20,9 @@ rmdir $MOUNTPT/foo
 ln -s garbage_string:123 $MOUNTPT/bar
 rm $MOUNTPT/bar
 aurteardown
+if [ $? -ne 0 ]; then
+    echo "Failed to tear down Aurora"
+    exit 1
+fi
 
 exit 0
